Refs #11612. Fix header file inclusion
